refactor: define useReferences
